[0.4.14][BTRFS] Workaround unrecognized device CORE-17388
Avoids the wizard for unrecognized devices to pop
up after each and every reboot when BTRFS is used.
Also slightly 'improves' the displaying in device manager.
Both regressed by 0.4.14-dev-1535-g 194ea909fdff60294ce2d5434cbded63f5377754

This is no proper solution, but more or less the same
that we did for 0.4.13 release already.

The proper fix will come in 0.4.15-dev'ing with
the storage-stack-rewrite and does not require those
4 lines to be added in the btrfs.inf anymore.